Summary: changed the utm converter (it is not fixed on this commit), fixed bugs related to points uploading service. Updated bootstrap version

+++ MapWebSite/MapWebSite.Tests/Database/CheckTest.cs
@@ -52,7 +52,7 @@ namespace MapWebSite.Tests.Database
         public void UTMConverterTest()
         {
             Helper.UTMConverter converter = new Helper.UTMConverter();
-            var result = converter.ToLatLong(35, 'T', (double)632353.875m, (double)4919262.500m);
+            var result = converter.ToLatLong(35, 'T', 632353.875m, 4919262.500m);
         }
     }
 }
